Fix device-test blockers: start sync engine, login redirect, RLS helper
- app.dart: watch syncEngineProvider + attachmentUploadQueueProvider so the
  offline outbox drains at launch (created records now reach Supabase and the
  dashboard populates)
- sync_engine.dart / attachment_upload_queue.dart: initial drain() on start()
  in case the app launches already online
- app_router.dart: splash always routes; redirect logged-in users off auth
  routes to the dashboard
- 0006_rls_helpers.sql: app.has_min_rank param smallint -> integer (fixes
  42883 during schema apply)
- add supabase/apply_all.sql (reset preamble + all 15 migrations + seed) and
  supabase/bootstrap_admin.sql (first company + Administrator)

+++ supabase/migrations/0006_rls_helpers.sql
@@ -43,8 +43,10 @@ as $$
     and ur.company_id = (select company_id from public.user_profiles where user_id = auth.uid());
 $$;
 
--- Convenience threshold check.
-create or replace function app.has_min_rank(min_rank smallint)
+-- Convenience threshold check. Parameter is `integer` so callers can pass plain
+-- integer literals (e.g. app.has_min_rank(5)) — Postgres won't implicitly narrow
+-- integer -> smallint during function resolution.
+create or replace function app.has_min_rank(min_rank integer)
 returns boolean
 language sql stable security definer set search_path = public
 as $$
